:recycel: refactor: progressRate 타입 수정

+++ src/main/java/com/saisai/domain/ride/dto/response/RecentRideInfoRes.java
@@ -10,7 +10,7 @@ public record RecentRideInfoRes(
     String sigun,
     String courseImageUrl,
     Double distance,
-    Double progressRate,
+    Integer progressRate,
     LocalDate recentRideAt
 ) {
